// EzCatalog.SPA/ezcatalog-spa/src/app/core/api/generated/model/add-product-request.ts
/**
 * EzCatalog API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddProductRequest { 
    sku: string;
    name: string;
    priceAmount: number;
    priceCurrency: AddProductRequest.PriceCurrencyEnum;
}
export namespace AddProductRequest {
    export const PriceCurrencyEnum = {
        Pln: 'PLN',
        Eur: 'EUR',
        Usd: 'USD',
    } as const;
    export type PriceCurrencyEnum = typeof PriceCurrencyEnum[keyof typeof PriceCurrencyEnum];
}